aws/dynamodb/sample1/src/main/java/com/mobicare/dynamodb/sample1/model/User.java
package com.mobicare.dynamodb.sample1.model;

import java.util.Date;

import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBHashKey;
import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBTable;

@DynamoDBTable(tableName = User.TABLE_NAME)
public class User {

	public static final String TABLE_NAME = "user";

	@DynamoDBHashKey(attributeName = "id")
	private String email;

	private String name;

	private Date date;

	public String getEmail() {
		return email;
	}

	public void setEmail(final String email) {
		this.email = email;
	}

	public String getName() {
		return name;
	}

	public void setName(final String name) {
		this.name = name;
	}

	public Date getDate() {
		return date;
	}

	public void setDate(final Date date) {
		this.date = date;
	}

}
